refactor: separate normalizeToRange function as utils
